requirements.txt
Django>=5.2
gunicorn
google-cloud-firestore
django-storages[google]
django-allauth[socialaccount]
requests
Pillow
